+++ src/tests/Timer_tests.cpp
@@ -43,7 +43,7 @@ TEST_CASE("Timer test TIMA increase")
     Timer t(bus);
 
     SUBCASE("Increase with ClockSelect::N16") {
-        t.writeTAC(Timer::TACTimerEnableMask | Timer::ClockSelect::N16);
+        t.writeTAC(Timer::TACTimerEnableMask | (uint8_t)Timer::ClockSelect::N16);
 
         t.step(4);
         CHECK(t.readTIMA() == 1);
@@ -55,7 +55,7 @@ TEST_CASE("Timer test TIMA increase")
     }
 
     SUBCASE("Increase with ClockSelect::N64") {
-        t.writeTAC(Timer::TACTimerEnableMask | Timer::ClockSelect::N64);
+        t.writeTAC(Timer::TACTimerEnableMask | (uint8_t)Timer::ClockSelect::N64);
 
         t.step(16);
         CHECK(t.readTIMA() == 1);
@@ -67,7 +67,7 @@ TEST_CASE("Timer test TIMA increase")
     }
 
     SUBCASE("Increase with ClockSelect::N256") {
-        t.writeTAC(Timer::TACTimerEnableMask | Timer::ClockSelect::N256);
+        t.writeTAC(Timer::TACTimerEnableMask | (uint8_t)Timer::ClockSelect::N256);
 
         t.step(64);
         CHECK(t.readTIMA() == 1);
@@ -79,7 +79,7 @@ TEST_CASE("Timer test TIMA increase")
     }
 
     SUBCASE("Increase with ClockSelect::N1024") {
-        t.writeTAC(Timer::TACTimerEnableMask | Timer::ClockSelect::N1024);
+        t.writeTAC(Timer::TACTimerEnableMask | (uint8_t)Timer::ClockSelect::N1024);
 
         t.step(256);
         CHECK(t.readTIMA() == 1);
@@ -98,7 +98,7 @@ TEST_CASE("Timer test TIMA overflow and interrupt generation")
     
     auto timerIrqMask = Irqs::mask(Irqs::Type::Timer);
 
-    t.writeTAC(Timer::TACTimerEnableMask | Timer::ClockSelect::N16);
+    t.writeTAC(Timer::TACTimerEnableMask | (uint8_t)Timer::ClockSelect::N16);
     
     t.writeTIMA(0xFE);
     // TIMA start from FE so we need 16 + 16 clock cycles (4 + 4 machine cycles) to overflow
@@ -130,7 +130,7 @@ TEST_CASE("Timer test changing subclock while timer is running")
     TestBus bus;
     Timer t(bus);
 
-    t.writeTAC(Timer::TACTimerEnableMask | Timer::ClockSelect::N16);
+    t.writeTAC(Timer::TACTimerEnableMask | (uint8_t)Timer::ClockSelect::N16);
 
     // 4 machine cycles = 16 clock cycles --> TIMA must become 1 (and subcounter 0)
     t.step(4);
@@ -142,7 +142,7 @@ TEST_CASE("Timer test changing subclock while timer is running")
     CHECK(t.getTimaSubcounter() == 8);
 
     // change subclock to N64
-    t.writeTAC(Timer::TACTimerEnableMask | Timer::ClockSelect::N64);
+    t.writeTAC(Timer::TACTimerEnableMask | (uint8_t)Timer::ClockSelect::N64);
 
     // 14 machine cycles = 56 clock cycles, TIMA must become 2 (and subcounter 0)
     t.step(14);
@@ -156,7 +156,7 @@ TEST_CASE("Timer test starting, stopping and restarting the timer")
     TestBus bus;
     Timer t(bus);
 
-    t.writeTAC(Timer::TACTimerEnableMask | Timer::ClockSelect::N16);
+    t.writeTAC(Timer::TACTimerEnableMask | (uint8_t)Timer::ClockSelect::N16);
 
     // 2 machine cycles = 8 clock cycles --> TIMA must still be 0 (and subcounter 8)
     t.step(2);
@@ -164,7 +164,7 @@ TEST_CASE("Timer test starting, stopping and restarting the timer")
     CHECK(t.getTimaSubcounter() == 8);
 
     // disable timer
-    t.writeTAC(Timer::ClockSelect::N16);
+    t.writeTAC((uint8_t)Timer::ClockSelect::N16);
     
     // 4 machine cycles = 16 clock cycles but timer disabled --> TIMA must still be 0 (and subcounter 8)
     t.step(4);
@@ -172,7 +172,7 @@ TEST_CASE("Timer test starting, stopping and restarting the timer")
     CHECK(t.getTimaSubcounter() == 8);
 
     // re-enable timer 
-    t.writeTAC(Timer::TACTimerEnableMask | Timer::ClockSelect::N16);
+    t.writeTAC(Timer::TACTimerEnableMask | (uint8_t)Timer::ClockSelect::N16);
 
     // 4 machine cycles = 16 clock cycles --> TIMA must become 1 (and subcounter 0 because it resets on timer enable)
     t.step(4);
